refactor(models): split model.py by lifetime, not by size
`model.py` held two things with different lifetimes: the canonical `Model`
composition root, which the plane keeps evolving, and the one-cycle
compatibility surface, which is scheduled to die. Kept in one file, removing
the compat layer is surgery on a live module; split out, it is a file deletion.

  model.py           1260 -> 664
  _legacy_bridge.py       476   task kwargs <-> Request/Response, ModelOutput
  _legacy_binding.py      170   RuntimeBindingPlan for the bare wrappers

The split holds because the bridge was already free-function shaped, as the
review measured: `_kwargs_to_request` (212 lines) touched only `self.dialect_id`
and the static `self._validate_n`, and `_response_to_model_output` only
`self.meta()`. Those three become parameters; nothing else changed.

Direction is now one-way, which is the property that makes the eventual delete
cheap: `_legacy_bridge` imports `deployment` and `ir` and nothing from `model`,
while `model` imports six names from it. `_legacy_binding` is reachable only
from `chat_model`/`gen_model`, so it goes when the wrappers go. Still zero
runtime import cycles across the 27 modules.

`uuid4()`'s non-deterministic binding fingerprint now sits alone in
`_legacy_binding`, where it is visible and removable, instead of buried
mid-file.

No behaviour change, and this is checked rather than asserted: every relocated
definition is byte-identical to its `main` source, or AST-identical once the
four renames this commit documents are applied (`_named_json_value`,
`_validate_n`, `_kwargs_to_request` and `_response_to_model_output` lose the
leading underscore on leaving the class), and the three self-to-parameter
conversions are AST-identical after the substitutions above.

That equivalence is measured against the current `main`, which matters because
#100 landed between the two and edited two of the definitions relocated here:
it narrowed `_named_json_value`'s sequence branch and `_kwargs_to_request`'s
`stop` branch to `list`/`tuple`. Both narrowings are carried into the new
homes, and the tests #100 added to guard them were confirmed to still fail when
either branch is widened back where it now lives -- three cases on the coercer,
two on `stop`.
